fix: users app, ratings app, orders app, products app 수정 -적절하지 않은 datafield type 변경 -users.models.py User class_email 속성 max_length 변경

+++ orders/models.py
@@ -3,8 +3,8 @@ from django.db import models
 class Order(models.Model):
     user            = models.ForeignKey('users.User', on_delete=models.CASCADE)
     order_status    = models.ForeignKey('OrderStatus', on_delete=models.PROTECT)
-    total_price     = models.IntegerField(null=True)
-    shipping_price  = models.IntegerField(null=True)
+    total_price     = models.DecimalField(max_digits=20, decimal_places=2, null=True)
+    shipping_price  = models.DecimalField(max_digits=20, decimal_places=2, null=True)
     order_date_time = models.DateTimeField(auto_now=True)
 
     class Meta:
@@ -43,7 +43,7 @@ class PaymentMethod(models.Model):
 
 class DiscountType(models.Model):
     name  = models.CharField(max_length=45)
-    rate  = models.IntegerField()
+    rate  = models.DecimalField(max_digits=10,decimal_places=2)
     order = models.ForeignKey('Order', on_delete=models.CASCADE)
 
     class Meta:
